fix(lint): 🚨 fixing sonarlint warnings
