+++ sonar-project.properties
@@ -4,4 +4,4 @@ sonar.projectKey=spa5k_quran
 # relative paths to source directories. More details and properties are described
 # in https://sonarcloud.io/documentation/project-administration/narrowing-the-focus/
 sonar.sources=.
-sonar.exclusions=./packages/scraper/**
+sonar.exclusions=packages/scraper/**/*
