+++ tanzil/src/features/index.ts
@@ -1,2 +1,3 @@
 export * from './mushaf';
 export * from './salah';
+export * from './surah';
