+++ tanzil/src/main.css
@@ -5,6 +5,4 @@
 * {
 	font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
 }
-@media (prefers-color-scheme: dark){
-  
-}
+
